Start and terminate presentation are now moved in lib/live as requested
