+++ lib/live/index.js
@@ -1,0 +1,3 @@
+module.exports = {
+	live:                require('./live'),
+}
