+++ routes/user/presentations/presentation/routes.js
@@ -1,4 +1,5 @@
 var handlers   = require('./handlers');
+var live   = require('../../../../lib/live').live;
 var logger     = require('logger-asq');
 
 module.exports.setUp = function setUp(app, middleware) {
@@ -13,10 +14,10 @@ module.exports.setUp = function setUp(app, middleware) {
     handlers.editPresentation);
 
   app.post('/:user/presentations/:presentationId/live',
-    middleware.isRouteOwner, handlers.startPresentation);
+    middleware.isRouteOwner, live.startPresentation);
 
   app.delete('/:user/presentations/:presentationId/live',
-    middleware.isRouteOwner, handlers.terminatePresentation);
+    middleware.isRouteOwner, live.terminatePresentation);
 
   app.get('/:user/presentations/:presentationId/live/:liveId', 
     middleware.authorizeLiveSession, handlers.livePresentation);
